Improve movement stability
Use navmesh boundary RTree to control velocity after RVO. If the velocity vector intersect the boundary, then project it to the boundary edge and use it as a new velocity. It will not prevent agents to stay at walkable area. When there are many agent in small area, then some of them goes through corners. But snapping to the navmesh solve this problem. It looks like velocity postprocessing (in rvo system) is too expensive for performance. But it increase the accuracy of the movement.