Implement search enemies for non-player entities
Each entity looks around in some radius and search entity from non-friend team. If it find someone, then add it to the separate list and add these values to all friend entities in the search radius.

Each entity has a main team (int32 id) and also a list of frient teams.  Entity is a friend if it either has the same main team or contained in the list. It's possible that A is a friend for B, but B is not friend for A.